requirements.txt
flask
gunicorn

numpy
pandas
scikit-learn
joblib

requests
beautifulsoup4
lxml
tldextract
urllib3

python-whois
python-dateutil
googlesearch-python

matplotlib
seaborn

requests
